+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [EmptyView] Added EmptyViewModel as a property
 - [ErrorView] is now public
 - [ErrorView] Added ErrorViewModel as a property
+- [BarcodeScanner][iOS] Made sure capture output runs on its own dispatcher.
 
 ## [26.0.0]
 - [BreakingChange][Camera] Renamed `Preview` to `CameraPreview`